+++ NEWS.md
@@ -1,5 +1,17 @@
 # rnote 0.2.0
 
+## New Features
+
+* Moved note file storage to default R data directory
+* Improved performance with RDS storage
+* Added function to initialise saving notes to project
+
+## Bug Fixes
+* Fixed issue with losing notes on update
+
+## Removed
+* backup.notes function removed
+
 # rnote 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.